pyproject.toml
[project]
name = "final-year-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "groq>=1.0.0",
    "instructor>=1.14.5",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-text-splitters>=1.1.1",
    "pinecone>=8.1.0",
    "psycopg2>=2.9.11",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.1",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.2.3",
]
